Shift into the sign bit as unsigned rather than as int
Shifting a 1 into bit 31 of a signed int is undefined behaviour, and the
engine does it in four places that all want the unsigned bit pattern and have
always got it: the netchan sequence words mask off and set the top bit to
carry the reliable flag, U_SCALE is the one entity-update flag that reaches
bit 31, the playerstate stat mask covers all 32 stats, and MSG_ReadLong
assembles its top byte from a value that exceeds INT_MAX once the byte is
128 or more.

None of this changes what the server puts on the wire. GCC compiles the
netchan and playerstate writers to the same and/shl/or against the same
0x7fffffff, and the only difference anywhere is which registers and stack
slots it picks in Netchan_Transmit and MSG_ReadLong.

Found by the first run of an instrumented kpded2 - the sanitizer had only
ever covered the game library, which never sees a packet the engine has not
already unpacked. Nine of the sixteen findings were this, and clearing them
is what lets the remaining one stand out.

+++ qcommon/common.c
@@ -1151,7 +1151,7 @@ int MSG_ReadLong (sizebuf_t *msg_read)
 		c = msg_read->data[msg_read->readcount]
 		+ (msg_read->data[msg_read->readcount+1]<<8)
 		+ (msg_read->data[msg_read->readcount+2]<<16)
-		+ (msg_read->data[msg_read->readcount+3]<<24);
+		+ ((uint32)msg_read->data[msg_read->readcount+3]<<24);
 	
 	msg_read->readcount += 4;
